ct fix7

+++ tasks/seq/agafeev_s_strassen_alg/src/strassen_seq.cpp
@@ -95,7 +95,7 @@ bool MultiplMatrixSequental::PreProcessingImpl() {
   auto* temp_ptr2 = reinterpret_cast<double*>(task_data->inputs[1]);
   second_input_.insert(second_input_.begin(), temp_ptr2,
                        temp_ptr2 + (task_data->inputs_count[2] * task_data->inputs_count[3]));
-  size_ = task_data->inputs_count[0];
+  size_ = static_cast<int>(task_data->inputs_count[0]);
 
   return true;
 }
